add(docs):增加英语页脚
